add remaining redispatch functionality
