+++ covalent_dispatcher/_core/__init__.py
@@ -18,5 +18,5 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
-from .data_manager import make_dispatch
+from .data_manager import make_derived_dispatch, make_dispatch
 from .dispatcher import cancel_workflow, run_dispatch
