+++ covalent_dispatcher/_core/dispatcher.py
@@ -144,6 +144,7 @@ async def _submit_task(result_object, node_id):
 
     # Get name of the node for the current task
     node_name = result_object.lattice.transport_graph.get_node_value(node_id, "name")
+    node_status = result_object.lattice.transport_graph.get_node_value(node_id, "status")
     app_log.debug(f"7A: Node name: {node_name} (run_planned_workflow).")
 
     # Handle parameter nodes
@@ -153,15 +154,27 @@ async def _submit_task(result_object, node_id):
         app_log.debug(f"7C: Node output: {output} (run_planned_workflow).")
         app_log.debug("8: Starting update node (run_planned_workflow).")
 
+        timestamp = datetime.now(timezone.utc)
         node_result = {
             "node_id": node_id,
-            "start_time": datetime.now(timezone.utc),
-            "end_time": datetime.now(timezone.utc),
+            "start_time": timestamp,
+            "end_time": timestamp,
             "status": Result.COMPLETED,
             "output": output,
         }
         await datasvc.update_node_result(result_object, node_result)
         app_log.debug("8A: Update node success (run_planned_workflow).")
+
+    elif node_status == Result.COMPLETED:
+        timestamp = datetime.now(timezone.utc)
+        node_result = {
+            "node_id": node_id,
+            "start_time": timestamp,
+            "end_time": timestamp,
+            "status": Result.COMPLETED,
+        }
+        await datasvc.update_node_result(result_object, node_result)
+        app_log.debug(f"Skipped completed node {node_id}.")
 
     else:
