+++ covalent_dispatcher/_service/app.py
@@ -28,125 +28,18 @@ from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 
 import covalent_dispatcher as dispatcher
-from covalent import lattice
 from covalent._results_manager.result import Result
 from covalent._shared_files import logger
-from covalent._shared_files.util_classes import Status
-from covalent._workflow.transport import TransportableObject
 
 from .._db.datastore import workflow_db
 from .._db.dispatchdb import DispatchDB
+from .._db.load import _result_from
 from .._db.models import Lattice
-from .._db.write_result_to_db import load_file
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
 
 router: APIRouter = APIRouter()
-
-
-def _result_from(lattice_record: Lattice) -> Result:
-    """Re-hydrate result object from the lattice record."""
-
-    function = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.function_filename
-    )
-    function_string = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.function_string_filename
-    )
-    function_docstring = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.docstring_filename
-    )
-    executor_data = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.executor_data_filename
-    )
-    workflow_executor_data = load_file(
-        storage_path=lattice_record.storage_path,
-        filename=lattice_record.workflow_executor_data_filename,
-    )
-    inputs = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.inputs_filename
-    )
-    named_args = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.named_args_filename
-    )
-    named_kwargs = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.named_kwargs_filename
-    )
-    error = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.error_filename
-    )
-    transport_graph = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.transport_graph_filename
-    )
-    output = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.results_filename
-    )
-    deps = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.deps_filename
-    )
-    call_before = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.call_before_filename
-    )
-    call_after = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.call_after_filename
-    )
-    cova_imports = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.cova_imports_filename
-    )
-    lattice_imports = load_file(
-        storage_path=lattice_record.storage_path, filename=lattice_record.lattice_imports_filename
-    )
-
-    name = lattice_record.name
-    executor = lattice_record.executor
-    workflow_executor = lattice_record.workflow_executor
-    num_nodes = lattice_record.electron_num
-
-    attributes = {
-        "workflow_function": function,
-        "workflow_function_string": function_string,
-        "__name__": name,
-        "__doc__": function_docstring,
-        "metadata": {
-            "executor": executor,
-            "executor_data": executor_data,
-            "workflow_executor": workflow_executor,
-            "workflow_executor_data": workflow_executor_data,
-            "deps": deps,
-            "call_before": call_before,
-            "call_after": call_after,
-        },
-        "args": inputs["args"],
-        "kwargs": inputs["kwargs"],
-        "named_args": named_args,
-        "named_kwargs": named_kwargs,
-        "transport_graph": transport_graph,
-        "cova_imports": cova_imports,
-        "lattice_imports": lattice_imports,
-        "post_processing": False,
-        "electron_outputs": {},
-    }
-
-    def dummy_function(x):
-        return x
-
-    lat = lattice(dummy_function)
-    lat.__dict__ = attributes
-
-    result = Result(
-        lat,
-        dispatch_id=lattice_record.dispatch_id,
-    )
-    result._root_dispatch_id = lattice_record.root_dispatch_id
-    result._status = Status(lattice_record.status)
-    result._error = error if error else None
-    result._inputs = inputs
-    result._start_time = lattice_record.started_at
-    result._end_time = lattice_record.completed_at
-    result._result = output if output is not None else TransportableObject(None)
-    result._num_nodes = num_nodes
-    return result
 
 
 @router.post("/submit")
@@ -166,8 +59,24 @@ async def submit(request: Request) -> UUID:
     data = await request.json()
     data = json.dumps(data).encode("utf-8")
 
-    dispatch_id = await dispatcher.run_dispatcher(data)
-    return dispatch_id
+    return await dispatcher.run_dispatcher(data)
+
+
+@router.post("/redispatch")
+async def redispatch(request: Request) -> str:
+
+    data = await request.json()
+    dispatch_id = data["dispatch_id"]
+    json_lattice = data["json_lattice"]
+    electron_updates = data["electron_updates"]
+    reuse_previous_results = data["reuse_previous_results"]
+
+    return await dispatcher.run_redispatch(
+        dispatch_id,
+        json_lattice,
+        electron_updates,
+        reuse_previous_results,
+    )
 
 
 @router.post("/cancel")
@@ -225,11 +134,10 @@ def get_result(
                 ).decode()
             return output
 
-        response = JSONResponse(
+        return JSONResponse(
             status_code=503,
             content={
                 "message": "Result not ready to read yet. Please wait for a couple of seconds."
             },
             headers={"Retry-After": "2"},
         )
-        return response
