+++ covalent_dispatcher/entry_point.py
@@ -52,6 +52,21 @@ async def run_dispatcher(json_lattice: str):
     return dispatch_id
 
 
+async def run_redispatch(
+    dispatch_id: str, json_lattice: str, electron_updates: dict, reuse_previous_results: bool
+):
+    from ._core import make_derived_dispatch, run_dispatch
+
+    redispatch_id = make_derived_dispatch(
+        dispatch_id, json_lattice, electron_updates, reuse_previous_results
+    )
+    run_dispatch(redispatch_id)
+
+    app_log.debug(f"Re-dispatching {dispatch_id} as {redispatch_id}")
+
+    return redispatch_id
+
+
 def cancel_running_dispatch(dispatch_id: str) -> None:
     """
     Cancels a running dispatch job.
